refactor: reorganize local data source structure
